[#809] Fix clippy warnings introduced with 1.88

+++ iceoryx2-cal/tests/shm_allocator_trait_tests.rs
@@ -135,7 +135,7 @@ mod shm_allocator {
         let layout = unsafe {
             Layout::from_size_align_unchecked(
                 CHUNK_SIZE,
-                (test.sut().max_alignment() + 1).next_power_of_two() as usize,
+                (test.sut().max_alignment() + 1).next_power_of_two(),
             )
         };
         let distance = unsafe { test.sut().allocate(layout) };
